설정 배경음악 효과음 연결

+++ FunhiPrototype2/FunhiPrototype2/DataModel.swift
@@ -12,8 +12,8 @@ import UIKit
 
 // 프라퍼티 설명:  언어 설정, 배경음, 효과음, 넘김 속도
 struct Setting {
-    var bgmVolume: Double
-    var effectVolume: Double
+    var bgmVolume: Float
+    var effectVolume: Float
     var textSpeed: Double
     var darkmode: Bool
